Move CommandBuffer into graphics/commands dir
