+++ src/RealEngine/graphics/commands/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(RealEngine PRIVATE
+    CommandBuffer.hpp           CommandBuffer.cpp
+)
