+++ src/RealEngine/graphics/commands/CommandBuffer.cpp
@@ -1,7 +1,7 @@
 ﻿/*!
  *  @author    Dubsky Tomas
  */
-#include <RealEngine/graphics/CommandBuffer.hpp>
+#include <RealEngine/graphics/commands/CommandBuffer.hpp>
 
 namespace re {
 
@@ -38,30 +38,6 @@ void CommandBuffer::submitToComputeQueue(
     const vk::Fence&                             signalFence /* = nullptr*/
 ) {
     computeQueue().submit2(submits, signalFence);
-}
-
-void CommandBuffer::beginDebugUtilsLabel(const char* label, glm::vec4 color /* = {}*/) const {
-    if constexpr (k_buildType == BuildType::Debug) {
-        m_cmdBuf.beginDebugUtilsLabelEXT(
-            vk::DebugUtilsLabelEXT{label, {color.r, color.g, color.b, color.a}},
-            dispatchLoaderDynamic()
-        );
-    }
-}
-
-void CommandBuffer::endDebugUtilsLabel() const {
-    if constexpr (k_buildType == BuildType::Debug) {
-        m_cmdBuf.endDebugUtilsLabelEXT(dispatchLoaderDynamic());
-    }
-}
-
-void CommandBuffer::insertDebugUtilsLabel(const char* label, glm::vec4 color /* = {}*/) const {
-    if constexpr (k_buildType == BuildType::Debug) {
-        m_cmdBuf.insertDebugUtilsLabelEXT(
-            vk::DebugUtilsLabelEXT{label, {color.r, color.g, color.b, color.a}},
-            dispatchLoaderDynamic()
-        );
-    }
 }
 
 void CommandBuffer::submitToGraphicsQueue(const vk::Fence& signalFence /* = nullptr*/) const {
